Add files via upload
In this project, I used data for the City of Toronto to visualize COVID data. Specifically, how the cases changed by day and the predominant outbreak type each week. Then I created 3 different maps. The first shows the percentage of 18-64 year olds that are classified as low income. The second outlines the number of COVID cases per 100,000 in the 140 neighborhoods of Toronto. Finally, the last one allocates a color to each neighborhood that simultaneously represents its case rate and low income percentage.